add basic sampling

+++ app.py
@@ -208,7 +208,7 @@ class app(base_app):
         if str(self.cfg['param']['algorithm']) == 'Visual Curvature' :
             command_args = command_args + ['-c', str(self.cfg['param']['scale']/100.0)]
         if str(self.cfg['param']['algorithm']) == 'Basic Sampling' :
-            command_args = command_args + ['-p', str(self.cfg['param']['scale']*3.0)]
+            command_args = command_args + ['-p', str(int(self.cfg['param']['scale']*3.0))]
             
         f = open(self.work_dir+"algoLog.txt", "a")
         cmd = self.runCommand(command_args, None, f)
